HW3/requirements.txt
fastapi~=0.115.12
uvicorn[standard]
sqlalchemy>=1.4  # поддержка AsyncSession
asyncpg          # драйвер PostgreSQL для async SQLAlchemy
redis[asyncio]>=4.2.0   # асинхронный клиент Redis
python-jose[cryptography]  # JWT
passlib[bcrypt]~=1.7.4
pydantic~=2.11.1
python-multipart
alembic~=1.15.2
jose
python-dotenv>=1.0.0
